% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acyRsa.R
\docType{package}
\name{acyRsa-package}
\alias{acyRsa}
\alias{acyRsa-package}
\title{acyRsa: Interface to ACyRSA's API BO}
\description{
Provides a convenient wrapper for consuming data from ACyRSA's BackOffice API.
}
\author{
\strong{Maintainer}: Augusto Hassel \email{augusto.hassel@gmail.com}

}
