% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_activos_aceptados}
\alias{acyrsa_activos_aceptados}
\title{Collateral List}
\usage{
acyrsa_activos_aceptados(connection, date = Sys.Date())
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal acyRsaConnection class object}

\item{date}{String. Date with format '\%Y-\%m-\%d'.}
}
\value{
A tibble.
}
\description{
Brings a list of the products thar are currently accepted as collateral.
}
