% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_connection}
\alias{acyrsa_connection}
\title{Create acyRsa Connection Object}
\usage{
acyrsa_connection(token, base_url)
}
\arguments{
\item{token}{String. **Mandaroty** Obtained with \code{\link{acyrsa_login}}}

\item{base_url}{String. **Mandaroty** URL given by  \code{\link{acyrsa_login}} or known by the client.}
}
\value{
Creates a 'acyrsa_connection' S4 Object
}
\description{
\code{acyrsa_connection} creates a New Connection Object. It must be created manually when using de API with the token authentication.
}
\note{
You can use accesors to get information about the Object by using:
\itemize{
\item \code{token(conn)}
\item \code{base_url(conn)}
}
}
\examples{
\dontrun{conn <- acyrsa_connection(token = "1234", base_url = "https://api.anywhereportfolio.com.ar/")}
}
