% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_cotizaciones}
\alias{acyrsa_cotizaciones}
\title{Market Data}
\usage{
acyrsa_cotizaciones(
  connection,
  entry_type,
  date,
  Symbol,
  CFICode,
  MarketID,
  MarketSegmentID,
  SecurityGroup,
  SecurityType,
  SecurityIDSource
)
}
\arguments{
\item{connection}{S4. **Mandaroty** Formal acyRsaConnection class object}

\item{entry_type}{String. **Mandaroty** Vector of one or many values. See allowed values:
\itemize{
\item **2** = Tick By Tick
\item **3** = Index Value
\item **5** = Clossing Prices
\item **6** = Settlement Price
\item **B** = Trade Volume
\item **C** = Open Interest
\item **D** = Composite Underlying Price
}}

\item{date}{String. **Mandaroty** Clearing Business Day with format '\%Y-\%m-\%d'.}

\item{Symbol}{String. Contract Symbol.}

\item{CFICode}{String.}

\item{MarketID}{String. See allowed values:
\itemize{
\item **ROFX** = Matba Rofex
\item **XMAB** = MAE
}}

\item{MarketSegmentID}{String. See allowed values:
\itemize{
\item **DDF** = Financial
\item **DDA** = Agricultural
\item **DUAL** = Others
}}

\item{SecurityGroup}{String.}

\item{SecurityType}{String. See allowed values:
\itemize{
\item **FUT** = Future
\item **OPT** = Option
\item **FXSPOT** = FX Spot
\item **CFD** = Contract for differences
\item **CS** = Common Stock
\item **MF** = Mutual Fund
\item **SECLOAN** = Security Loan
\item **CD** = Certificate of Deposit
\item **GO** = General Obligation Bonds
\item **FXNDF** = Non-deliverable forward
\item **TD** = Time Deposit
}}

\item{SecurityIDSource}{String. See allowed values:
\itemize{
\item **4** = ISIN Number
}}
}
\value{
Data Frame
}
\description{
\code{acyrsa_cotizaciones} method allows to access Market Data from ACyRSA
}
\examples{
\dontrun{acyrsa_cotizaciones(connection = conn, entry_type = 6, date = "2020-04-17", SecurityType = "FUT")}
}
