% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_cuentas}
\alias{acyrsa_cuentas}
\title{Depositary Account List}
\usage{
acyrsa_cuentas(connection, market_account = TRUE)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal acyRsaConnection class object}
}
\value{
A tibble.
}
\description{
Brings a list with information of the accounts owned by the broker.
}
