% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_garantias_integradas}
\alias{acyrsa_garantias_integradas}
\title{Integrated Guarantees}
\usage{
acyrsa_garantias_integradas(connection, cim, alyc, date = Sys.Date())
}
\arguments{
\item{connection}{S4. Formal acyRsaConnection class object}

\item{cim}{Integer. Compensation Account Code}

\item{alyc}{Integer.Clearing Member Code}

\item{date}{String. Date with format '\%Y-\%m-\%d'.}
}
\value{
Data Frame
}
\description{
\code{acyrsa_garantias_integradas} method it's used to query the list of integrated garantees from our clients in ACyRSA.
}
