% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_horario_procesos}
\alias{acyrsa_horario_procesos}
\title{Closing Proce}
\usage{
acyrsa_horario_procesos(connection, date = Sys.Date(), type = NULL)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal acyRsaConnection class object}

\item{date}{String. Date with format '\%Y-\%m-\%d'.}

\item{type}{Integer. Type of process. If 'NULL' (default) is all.
\itemize{
\item \strong{1} - 'Pasaje a definitivo'
\item \strong{2} - 'Calculo de margenes y garantias'
\item \strong{11} - 'Actualizacion de activos'
\item \strong{16} - 'Facturacion de comisiones, bonificaciones y servicios'
\item \strong{19} - 'Publicacion de indice accionario'
\item \strong{21} - 'Movimientos de custodia'
\item \strong{22} - 'Publicacion del porfolio'
\item \strong{23} - 'Publicacion de ajustes'
}}
}
\value{
A tibble.
}
\description{
List all processes and it time stamp.
}
