% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_margenes}
\alias{acyrsa_margenes}
\title{Margin Requirement Report}
\usage{
acyrsa_margenes(connection, date = Sys.Date())
}
\arguments{
\item{connection}{S4. **Mandaroty** Formal acyRsaConnection class object}

\item{date}{String. Date with format '\%Y-\%m-\%d'.}
}
\value{
Data Frame
}
\description{
\code{acsa_margenes} method it's used to query the required margins for a certain date.
}
