% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\docType{class}
\name{acyRsaConnection-class}
\alias{acyRsaConnection-class}
\title{Connection Class: acyRsaConnection}
\description{
Creates a acyRsa connection object
}
\section{Slots}{

\describe{
\item{\code{token}}{character. Obtained from login method}

\item{\code{base_url}}{character. Connected environment}

\item{\code{valid_until}}{character. Log-in date. Valid for a day.}

\item{\code{agent}}{character. User Agent to pass to the API. Format: 'acyRsa-<environment>-user_name'}

\item{\code{user_name}}{character. User Name.}
}}

