% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{acyRsaConnection-method}
\alias{acyRsaConnection-method}
\alias{token}
\alias{base_url}
\alias{valid_until}
\alias{agent}
\alias{user_name}
\alias{token,acyRsaConnection-method}
\alias{base_url,acyRsaConnection-method}
\alias{valid_until,acyRsaConnection-method}
\alias{agent,acyRsaConnection-method}
\alias{user_name,acyRsaConnection-method}
\alias{show,acyRsaConnection-method}
\title{acyRsaConnection-method}
\usage{
token(x)

base_url(x)

valid_until(x)

agent(x)

user_name(x)

\S4method{token}{acyRsaConnection}(x)

\S4method{base_url}{acyRsaConnection}(x)

\S4method{valid_until}{acyRsaConnection}(x)

\S4method{agent}{acyRsaConnection}(x)

\S4method{user_name}{acyRsaConnection}(x)

\S4method{show}{acyRsaConnection}(object)
}
\arguments{
\item{x}{S4 Class. acyRsaConnection object}

\item{object}{S4 Class. acyRsaConnection object}
}
\description{
acyRsaConnection-method
}
