% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyRsaConnection}
\alias{acyRsaConnection}
\title{New acyRsa Connection}
\usage{
acyRsaConnection(token, base_url)
}
\arguments{
\item{token}{String. Obtained with \code{\link{acyrsa_login}}}

\item{base_url}{String. URL given by  \code{\link{acyrsa_login}} or known by the client.}
}
\value{
Creates a 'acyRsaConnection' S4 Object
}
\description{
\code{acyRsaConnection} creates a New Connection Object. It must be created manually when using de API with the token authentication.
}
\examples{
\dontrun{conn <- acyRsaConnection(token = "1234", base_url = "https://api.anywhereportfolio.com.ar/")}
}
