% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_connection}
\alias{acyrsa_connection}
\title{cyRsaConnection Object}
\usage{
acyrsa_connection(token, base_url)
}
\arguments{
\item{token}{String. \strong{Mandatory} Obtained with \code{\link{acyrsa_login}}}

\item{base_url}{String. \strong{Mandatory} URL given to \code{\link{acyrsa_login}} to initiate the connection.}
}
\value{
Creates an 'acyRsaConnection' S4 Object
}
\description{
`acyrsa_connection()` creates a New Connection Object. It should be created manually when saving the token outside the application.
}
\section{Accesors}{

You can use accesors to get information about the Object by using:
\itemize{
\item \code{token(conn)}
\item \code{base_url(conn)}
\item \code{valid_until(conn)}
}
}

\examples{
\dontrun{
conn <- acyrsa_connection(token = "1234", base_url = "https://api.anywhereportfolio.com.ar/")
}
}
\seealso{
Other connection functions: 
\code{\link{acyrsa_login}()}
}
\concept{connection functions}
