% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_cotizaciones}
\alias{acyrsa_cotizaciones}
\title{Market Data}
\usage{
acyrsa_cotizaciones(
  connection,
  entry_type,
  date,
  Symbol,
  CFICode,
  MarketID,
  MarketSegmentID,
  SecurityGroup,
  SecurityType,
  SecurityIDSource
)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal acyRsaConnection class object}

\item{entry_type}{String. \strong{Mandatory} Vector of one or many values. See allowed values:
\itemize{
\item \strong{2} = Tick By Tick
\item \strong{3} = Index Value
\item \strong{5} = Clossing Prices
\item \strong{6} = Settlement Price
\item \strong{B} = Trade Volume
\item \strong{C} = Open Interest
\item \strong{D} = Composite Underlying Price
}}

\item{date}{String. \strong{Mandatory} Clearing Business Day with format '\%Y-\%m-\%d'.}

\item{Symbol}{String. Contract Symbol.}

\item{CFICode}{String.}

\item{MarketID}{String. See allowed values:
\itemize{
\item \strong{ROFX} = Matba Rofex
\item \strong{XMAB} = MAE
}}

\item{MarketSegmentID}{String. See allowed values:
\itemize{
\item \strong{DDF} = Financial
\item \strong{DDA} = Agricultural
\item \strong{DUAL} = Others
}}

\item{SecurityGroup}{String.}

\item{SecurityType}{String. See allowed values:
\itemize{
\item \strong{FUT} = Future
\item \strong{OPT} = Option
\item \strong{FXSPOT} = FX Spot
\item \strong{CFD} = Contract for differences
\item \strong{CS} = Common Stock
\item \strong{MF} = Mutual Fund
\item \strong{SECLOAN} = Security Loan
\item \strong{CD} = Certificate of Deposit
\item \strong{GO} = General Obligation Bonds
\item \strong{FXNDF} = Non-deliverable forward
\item \strong{TD} = Time Deposit
}}

\item{SecurityIDSource}{String. See allowed values:
\itemize{
\item \strong{4} = ISIN Number
}}
}
\value{
Data Frame
}
\description{
`acyrsa_cotizaciones()` it's used to access Market Data from ACyRSA
}
\examples{
\dontrun{
acyrsa_cotizaciones(connection = conn, entry_type = 6, date = "2020-04-17", SecurityType = "FUT")
}
}
