% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_login}
\alias{acyrsa_login}
\title{Log-in Method}
\usage{
acyrsa_login(user, pass, env)
}
\arguments{
\item{user}{String. \strong{Mandatory} User Name}

\item{pass}{String. \strong{Mandatory} Password}

\item{env}{String. \strong{Mandatory} Wich environment are you going to connect:
\itemize{
\item \strong{demoapi} - for testing
\item \strong{api} - for production
}}
}
\value{
Creates an 'acyRsaConnection' S4 Object with a token and a base_url.
}
\description{
`acyrsa_login()` it's used to Log-in and obtained a valid token tthat then should be used in all requests to the API.
}
\note{
The token is valid only for a day.
}
\section{Accesors}{

You can use accesors to get information about the Object by using:
\itemize{
\item \code{token(conn)}
\item \code{base_url(conn)}
\item \code{login_date_time(conn)}
\item \code{agent(conn)}
\item \code{user_name(conn)}
}
}

\examples{
\dontrun{
conn <- acyrsa_login(user = "asd", pass = "xxx", env = "api")
}
}
\concept{connection functions}
