% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{acyrsa_margenes}
\alias{acyrsa_margenes}
\title{Margin Requirement Report}
\usage{
acyrsa_margenes(connection, date = Sys.Date(), viewDetails = FALSE)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal acyRsaConnection class object}

\item{date}{String. Date with format '\%Y-\%m-\%d'.}

\item{viewDetails}{Logical. With or without breakdown.}
}
\value{
A tibble.
}
\description{
It is use to query the required margins for a certain date.
}
